requirements.txt
click
psycopg2-binary
flask
pytest
watchdog
python-dotenv
rich
pytest
pytest-mock